theatre-service/requirements.txt
fastapi==0.111.0
elasticsearch[async]==8.13.2
redis==5.0.4
pydantic-settings==2.8.0
aiohttp==3.11.13
backoff==2.2.1
python-jose[cryptography]==3.3.0